refactor(jules-cli): split main.py into domain specific files and commands
- Moved models and client endpoints into `session.py`, `activity.py`, and `source.py`
- Created `client/` package with `JulesClient` acting as facade for domain-specific clients
- Created `commands/` package grouping typer subcommands by domain
- Simplified `main.py` to be entrypoint only
- Updated tests and paths
