+++ src/python/jules_cli/jules_cli/client/__init__.py
@@ -1,0 +1,10 @@
+from jules_cli.client.activity import ActivityClient
+from jules_cli.client.session import SessionClient
+from jules_cli.client.source import SourceClient
+
+
+class JulesClient(ActivityClient, SessionClient, SourceClient):
+    """
+    Client for interacting with the Jules API.
+    See: https://developers.google.com/jules/api.
+    """
