+++ src/python/jules_cli/jules_cli/commands/__init__.py
@@ -1,0 +1,5 @@
+from jules_cli.commands.activity import activity_app
+from jules_cli.commands.session import session_app
+from jules_cli.commands.source import source_app
+
+__all__ = ["activity_app", "session_app", "source_app"]
