+++ src/python/jules_cli/tests/test_main.py
@@ -10,9 +10,12 @@ import typer
 from typer.testing import CliRunner
 from whenever import Instant
 
+from jules_cli.activity import Activity
 from jules_cli.config import JulesConfig
+from jules_cli.core import JulesContext
 from jules_cli.main import cli, get_api_key
-from jules_cli.models import Activity, GitHubRepo, JulesContext, Session, Source, SourceContext
+from jules_cli.session import Session
+from jules_cli.source import GitHubRepo, Source, SourceContext
 
 
 def test_get_api_key_ignores_env_var() -> None:
@@ -48,7 +51,7 @@ def test_cli_help_example_toml() -> None:
 
 def test_list_sessions_cli_override() -> None:
     runner = CliRunner()
-    with patch("jules_cli.main.JulesClient") as mock_client:
+    with patch("jules_cli.commands.session.JulesClient") as mock_client:
         # Mock context manager
         instance = mock_client.return_value
         instance.__aenter__ = AsyncMock(return_value=instance)
@@ -72,7 +75,7 @@ def test_list_sessions_cli_override() -> None:
 
 def test_list_sessions_json() -> None:
     runner = CliRunner()
-    with patch("jules_cli.main.JulesClient") as mock_client:
+    with patch("jules_cli.commands.session.JulesClient") as mock_client:
         instance = mock_client.return_value
         instance.__aenter__ = AsyncMock(return_value=instance)
         instance.__aexit__ = AsyncMock(return_value=None)
@@ -99,7 +102,7 @@ def test_list_sessions_json() -> None:
 
 def test_show_session_json() -> None:
     runner = CliRunner()
-    with patch("jules_cli.main.JulesClient") as mock_client:
+    with patch("jules_cli.commands.session.JulesClient") as mock_client:
         instance = mock_client.return_value
         instance.__aenter__ = AsyncMock(return_value=instance)
         instance.__aexit__ = AsyncMock(return_value=None)
@@ -125,12 +128,11 @@ def test_show_session_json() -> None:
         data = json.loads(result.output)
         assert data["id"] == "1"
         assert data["title"] == "Test Session"
-        assert data["activities"] == []
 
 
 def test_create_session_json() -> None:
     runner = CliRunner()
-    with patch("jules_cli.main.JulesClient") as mock_client:
+    with patch("jules_cli.commands.session.JulesClient") as mock_client:
         instance = mock_client.return_value
         instance.__aenter__ = AsyncMock(return_value=instance)
         instance.__aexit__ = AsyncMock(return_value=None)
@@ -167,7 +169,7 @@ def test_create_session_json() -> None:
 
 def test_message_session_with_message() -> None:
     runner = CliRunner()
-    with patch("jules_cli.main.JulesClient") as mock_client:
+    with patch("jules_cli.commands.session.JulesClient") as mock_client:
         instance = mock_client.return_value
         instance.__aenter__ = AsyncMock(return_value=instance)
         instance.__aexit__ = AsyncMock(return_value=None)
@@ -175,7 +177,7 @@ def test_message_session_with_message() -> None:
 
         result = runner.invoke(
             cli,
-            ["--api-key", "key", "session", "message", "sessions/1", "-m", "hello"],
+            ["--api-key", "key", "session", "message", "sessions/1", "hello"],
         )
 
         assert result.exit_code == 0
@@ -231,7 +233,7 @@ def test_cli_context_type() -> None:
 
 def test_approve_session_cli() -> None:
     runner = CliRunner()
-    with patch("jules_cli.main.JulesClient") as mock_client:
+    with patch("jules_cli.commands.session.JulesClient") as mock_client:
         instance = mock_client.return_value
         instance.__aenter__ = AsyncMock(return_value=instance)
         instance.__aexit__ = AsyncMock(return_value=None)
@@ -249,7 +251,7 @@ def test_approve_session_cli() -> None:
 
 def test_approve_session_cli_json() -> None:
     runner = CliRunner()
-    with patch("jules_cli.main.JulesClient") as mock_client:
+    with patch("jules_cli.commands.session.JulesClient") as mock_client:
         instance = mock_client.return_value
         instance.__aenter__ = AsyncMock(return_value=instance)
         instance.__aexit__ = AsyncMock(return_value=None)
@@ -268,7 +270,7 @@ def test_approve_session_cli_json() -> None:
 
 def test_activity_show_cli() -> None:
     runner = CliRunner()
-    with patch("jules_cli.main.JulesClient") as mock_client:
+    with patch("jules_cli.commands.activity.JulesClient") as mock_client:
         instance = mock_client.return_value
         instance.__aenter__ = AsyncMock(return_value=instance)
         instance.__aexit__ = AsyncMock(return_value=None)
@@ -288,7 +290,7 @@ def test_activity_show_cli() -> None:
 
 def test_activity_list_cli() -> None:
     runner = CliRunner()
-    with patch("jules_cli.main.JulesClient") as mock_client:
+    with patch("jules_cli.commands.activity.JulesClient") as mock_client:
         instance = mock_client.return_value
         instance.__aenter__ = AsyncMock(return_value=instance)
         instance.__aexit__ = AsyncMock(return_value=None)
@@ -308,7 +310,7 @@ def test_activity_list_cli() -> None:
 
 def test_source_show_cli() -> None:
     runner = CliRunner()
-    with patch("jules_cli.main.JulesClient") as mock_client:
+    with patch("jules_cli.commands.source.JulesClient") as mock_client:
         instance = mock_client.return_value
         instance.__aenter__ = AsyncMock(return_value=instance)
         instance.__aexit__ = AsyncMock(return_value=None)
@@ -326,7 +328,7 @@ def test_source_show_cli() -> None:
 
 def test_source_list_cli() -> None:
     runner = CliRunner()
-    with patch("jules_cli.main.JulesClient") as mock_client:
+    with patch("jules_cli.commands.source.JulesClient") as mock_client:
         instance = mock_client.return_value
         instance.__aenter__ = AsyncMock(return_value=instance)
         instance.__aexit__ = AsyncMock(return_value=None)
